Moved shared memory functions to utility
